+++ examples/penguins/app.py
@@ -94,10 +94,9 @@ def server(input: Inputs, output: Outputs, session: Session):
             return x.ui.value_box(
                 title,
                 count,
-                {"class_": "pt-1 pb-0"},
-                showcase=x.ui.bind_fill_role(
-                    ui.tags.img({"style": "object-fit:contain;"}, src=showcase_img),
-                    item=True,
+                {"class": "pt-1 pb-0"},
+                showcase=x.ui.as_fill_item(
+                    ui.tags.img({"style": "object-fit:contain;"}, src=showcase_img)
                 ),
                 theme_color=None,
                 style=f"background-color: {bgcol};",