+++ shiny/experimental/ui/_card.py
@@ -1,6 +1,6 @@
 from __future__ import annotations
 
-from typing import NamedTuple, Optional
+from typing import Optional
 
 from htmltools import Tag, TagAttrs, TagAttrValue, TagChild, css, div, tags
 
@@ -9,30 +9,8 @@ from ._card_full_screen import full_screen_toggle
 from ._card_item import CardItem, WrapperCallable, card_body, wrap_children_in_card
 from ._css import CssUnit, validate_css_unit
 from ._fill import bind_fill_role
-
-# class Page:
-#     x: Tag
-#     def __init__(
-#         self,
-#         x: Tag,
-#     ):
-#         self.x = x
-
-
-# class Fragment:
-#     x: Tag
-#     page: Page
-#     def __init__(
-#         self,
-#         x: Tag,
-#         page: Page,
-#     ):
-#         self.x = x
-#         self.page = page
-
-
-# def as_fragment(x: Tag, page: Page) -> Fragment:
-#     return Fragment(x=x, page=page)
+from ._htmldeps import card_dependency
+from ._utils import consolidate_attrs
 
 
 # A Bootstrap card component
@@ -95,98 +73,37 @@ def card(
     height: Optional[CssUnit] = None,
     max_height: Optional[CssUnit] = None,
     fill: bool = True,
-    class_: Optional[str] = None,  # Applies after `bind_fill_role()`
+    class_: Optional[str] = None,
     wrapper: WrapperCallable | None | MISSING_TYPE = MISSING,
     **kwargs: TagAttrValue,
 ) -> Tag:
     if isinstance(wrapper, MISSING_TYPE):
         wrapper = card_body
 
-    children, attrs = separate_args_into_children_and_attrs(*args)
+    attrs, children = consolidate_attrs(*args, class_=class_, **kwargs)
     children = wrap_children_in_card(*children, wrapper=wrapper)
 
     tag = div(
-        *children,
-        *attrs,
-        full_screen_toggle() if full_screen else None,
-        card_js_init(),
         {
             "class": "card bslib-card",
             "style": css(
                 height=validate_css_unit(height),
                 max_height=validate_css_unit(max_height),
             ),
+            "data-bslib-card-init": True,
         },
-        **kwargs,
+        *children,
+        attrs,
+        full_screen_toggle() if full_screen else None,
+        card_dependency(),
+        card_js_init(),
     )
 
-    tag = bind_fill_role(tag, container=True, item=fill)
-    # Give the user an opportunity to override the classes added by bind_fill_role()
-    if class_ is not None:
-        tag.add_class(class_)
-    return tag
-
-
-class ChildrenAndAttrs(NamedTuple):
-    children: list[TagChild | CardItem]
-    attrs: list[TagAttrs]
-
-
-def separate_args_into_children_and_attrs(
-    *args: TagChild | TagAttrs | CardItem,
-) -> ChildrenAndAttrs:
-    children: list[TagChild | CardItem] = []
-    attrs: list[TagAttrs] = []
-
-    for arg in args:
-        if isinstance(arg, dict):
-            attrs.append(arg)
-        else:
-            children.append(arg)
-
-    return ChildrenAndAttrs(children, attrs)
+    return bind_fill_role(tag, container=True, item=fill)
 
 
 def card_js_init() -> Tag:
     return tags.script(
-        {"data-bslib-card-needs-init": True},
-        """\
-      var thisScript = document.querySelector('script[data-bslib-card-needs-init]');
-      if (!thisScript) throw new Error('Failed to register card() resize observer');
-
-      thisScript.removeAttribute('data-bslib-card-needs-init');
-
-      var card = $(thisScript).parents('.card').last();
-      if (!card) throw new Error('Failed to register card() resize observer');
-
-      // Let Shiny know to trigger resize when the card size changes
-      // TODO: shiny could/should do this itself (rstudio/shiny#3682)
-      var resizeEvent = window.document.createEvent('UIEvents');
-      resizeEvent.initUIEvent('resize', true, false, window, 0);
-      var ro = new ResizeObserver(() => { window.dispatchEvent(resizeEvent); });
-      ro.observe(card[0]);
-
-      // Enable tooltips (for the expand icon)
-      var tooltipList = card[0].querySelectorAll('[data-bs-toggle=\"tooltip\"]');
-      tooltipList.forEach(function(x) { new bootstrap.Tooltip(x); });
-
-      // In some complex fill-based layouts with multiple outputs (e.g., plotly),
-      // shiny initializes with the correct sizing, but in-between the 1st and last
-      // renderValue(), the size of the output containers can change, meaning every
-      // output but the 1st gets initialized with the wrong size during their
-      // renderValue(); and then after the render phase, shiny won't know trigger a
-      // resize since all the widgets will return to their original size
-      // (and thus, Shiny thinks there isn't any resizing to do).
-      // We workaround that situation by manually triggering a resize on the binding
-      // when the output container changes (this way, if the size is different during
-      // the render phase, Shiny will know about it)
-      $(document).on('shiny:value', function(x) {
-        var el = x.binding.el;
-        if (card[0].contains(el) && !$(el).data('bslib-output-observer')) {
-          var roo = new ResizeObserver(x.binding.onResize);
-          roo.observe(el);
-          $(el).data('bslib-output-observer', true);
-        }
-      });
-    """,
+        {"data-bslib-card-init": True},
+        "window.bslib.Card.initializeAllCards();",
     )