+++ shiny/experimental/ui/_css.py
@@ -1,11 +1,8 @@
 from __future__ import annotations
 
-import numbers
 from typing import Union, overload
 
 CssUnit = Union[
-    # TODO: pylance really doesn't like `numbers.Number`.
-    #       Instead, use `int` and `float`
     int,
     float,
     str,
@@ -44,13 +41,7 @@ def validate_css_unit(value: CssUnit) -> str:
 def validate_css_unit(value: None | CssUnit) -> None | str:
     # TODO-future: Actually validate. Or don't validate, but then change
     # the function name to to_css_unit() or something.
-    # TODO-future: pylance can't figure out if an `int` or `float` is a `numbers.Number` (which
-    # is it). For now, use the extra types
-    if (
-        isinstance(value, numbers.Number)
-        or isinstance(value, float)
-        or isinstance(value, int)
-    ):
+    if isinstance(value, (float, int)):
         # Explicit check for 0 because floats may format to have many decimals.
         if value == 0:
             return "0"