+++ shiny/experimental/ui/_htmldeps.py
@@ -4,30 +4,33 @@ from pathlib import PurePath
 
 from htmltools import HTMLDependency
 
-from shiny import __version__ as shiny_package_version
+from ..._versions import bslib as bslib_version
+from ..._versions import htmltools as htmltools_version
 
-ex_www_path = PurePath(__file__).parent.parent / "www"
+x_www = PurePath(__file__).parent.parent / "www"
+x_components_path = x_www / "bslib" / "components"
+x_fill_path = x_www / "htmltools" / "fill"
 
 
-def card_full_screen_dep() -> HTMLDependency:
+def card_dependency() -> HTMLDependency:
     return HTMLDependency(
-        name="bslib-card-full-screen",
-        version=shiny_package_version,
+        name="bslib-card",
+        version=bslib_version,
         source={
             "package": "shiny",
-            "subdir": str(ex_www_path),
+            "subdir": str(x_components_path),
         },
-        script={"src": "card-full-screen.js"},
+        script={"src": "card.min.js"},
     )
 
 
-def fill_dependencies() -> HTMLDependency:
+def fill_dependency() -> HTMLDependency:
     return HTMLDependency(
         "htmltools-fill",
-        "0.0.0.0",
+        htmltools_version,
         source={
             "package": "shiny",
-            "subdir": str(ex_www_path),
+            "subdir": str(x_fill_path),
         },
         stylesheet={"href": "fill.css"},
     )
@@ -35,11 +38,23 @@ def fill_dependencies() -> HTMLDependency:
 
 def sidebar_dependency() -> HTMLDependency:
     return HTMLDependency(
-        "bslib-sidebar-x",
-        "0.0.0",
+        "bslib-sidebar",
+        bslib_version,
         source={
             "package": "shiny",
-            "subdir": str(ex_www_path / "sidebar"),
+            "subdir": str(x_components_path),
         },
         script={"src": "sidebar.min.js"},
+    )
+
+
+def accordion_dependency() -> HTMLDependency:
+    return HTMLDependency(
+        "bslib-accordion",
+        version=bslib_version,
+        source={
+            "package": "shiny",
+            "subdir": str(x_components_path),
+        },
+        script={"src": "accordion.min.js"},
     )